requirements.txt
boto3==1.40.45
faiss-cpu==1.12.0
google-cloud-vision==3.10.0
huggingface-hub==0.34.3
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
numpy==2.1.2
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
sentence-transformers==5.0.0
streamlit==1.49.1
tensorflow==2.19.0
tokenizers==0.21.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
transformers==4.54.0
